requirements.txt
 streamlit 
 pandas 
 google-api-python-client 
 mysql-connector-python 
 sqlalchemy 
 python-dotenv
 pymongo
 isodate

 -e .
 